Adds layout for generateMaze test
Generates mazes of random side lengths and tests to see whether
the mazes are "valid". They are valid if they have non_adjacent
entrances and exits and there is a path from the entrance to the
exit. The testing function is not yet working as two helper
functions need to be implemented.