Deletes a function and renames another
Deletes the buildBranches function which is no longer necessary.
It was originally envisioned that a solution path would first
be built and then branches would be built off that path. Instead
all the path area is built within one function and for that
reason it is renamed to buildAllPath, instead of
buildSolutionPath.